refactor: simplify link handling in Toolbar and LinkPopup components
